Move firewall package under pkg/utils
