+++ agent/store.go
@@ -3,7 +3,7 @@ package agent
 import (
 	"github.com/golang/glog"
 	"github.com/romana/core/common"
-	"github.com/romana/core/firewall"
+	"github.com/romana/core/pkg/util/firewall"
 	"sync"
 )
